Added grid and vector type check

+++ src/Grid/comb.hpp
@@ -169,7 +169,7 @@ struct comb
 	{
 		int zero = 0;
 
-		for (int i = 0 ; i < dim ; i++)
+		for (size_t i = 0 ; i < dim ; i++)
 		{
 			if (c[i] == 0) zero++;
 		}
